% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autores.R
\name{add_authors}
\alias{add_authors}
\title{Añadir autores de publicaciones}
\usage{
add_authors(aut.text = NULL, aut.bib = NULL, aut.sep = "; ", pause = 3)
}
\arguments{
\item{aut.text}{Cadena de texto conteniendo el nombre de uno o varios autores}

\item{aut.bib}{Referencia en formato BibTeX (aún no implementado)}

\item{aut.sep}{Caracter(es) que separa cada autor en \code{aut_text}}

\item{pause}{Tiempo (en segundos) que dura el nombre de cada autor en el portapapeles
(para que dé tiempo a pegar cada uno en la aplicación de la ANECA)}
}
\value{
Los autores son pegados al portapapeles
}
\description{
Añadir autores de publicaciones
}
\examples{
\dontrun{
add_authors("Quintero, Elena; Rodriguez-Sanchez, Francisco; Jordano, Pedro")
}
}
