% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANECAtools-package.R
\docType{package}
\name{ANECAtools-package}
\alias{ANECAtools}
\alias{ANECAtools-package}
\title{ANECAtools: Facilitando el proceso de acreditación en la ANECA}
\description{
Herramientas para facilitar el proceso de solicitud de acreditación de profesorado en la ANECA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/ANECAtools}
  \item \url{https://pakillo.github.io/ANECAtools/}
  \item Report bugs at \url{https://github.com/Pakillo/ANECAtools/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodríguez-Sánchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
