% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.R
\name{fecha_formatear}
\alias{fecha_formatear}
\title{Formatear fecha}
\usage{
fecha_formatear(fecha = NULL)
}
\arguments{
\item{fecha}{Fechas en formato "yyyy-mm-dd"}
}
\value{
Fechas en formato "dd/mm/yyyy"
}
\description{
Formatear fecha en el formato requerido por la ANECA (dd/mm/yyyy).
}
\examples{
fechas <- c("2024-02-28", "2023-12-25")
fecha_formatear(fechas)
}
