% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introducir_campos.R
\name{introducir_campos}
\alias{introducir_campos}
\title{Introducir campos de cualquier data frame}
\usage{
introducir_campos(df = NULL, pausa = 4)
}
\arguments{
\item{df}{Data frame con distintos campos a introducir en la aplicación de la ANECA.
Este data frame puede leerse en R p. ej.
a partir de un fichero CSV (usando \code{\link[readr:read_delim]{readr::read_csv()}})
o un fichero Excel (usando \code{\link[readxl:read_excel]{readxl::read_excel()}}).
Aunque no es estrictamente necesario, idealmente el data frame debe contener
los mismos campos (y en el mismo orden) requeridos por la aplicación de la ANECA.
El nombre de las columnas puede ser diferente al de la aplicación.}

\item{pausa}{Tiempo (en segundos) que dura cada campo en el portapapeles
(para que dé tiempo a pegar cada uno en la aplicación de la ANECA)}
}
\value{
Esta función va copiando los distintos campos al portapapeles para
poder pegarlos en la aplicación de la ANECA.
}
\description{
Esta función sirve para introducir información de los distintos campos requeridos
en cualquier sección de la aplicación ANECA (proyectos, congresos, etc).
La función simplemente lee el data frame y va pegando cada campo al portapapeles.
Para introducir publicaciones, debe utilizarse \code{\link[=introducir_publicaciones]{introducir_publicaciones()}}.
}
