% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicaciones.R
\name{introducir_publicaciones}
\alias{introducir_publicaciones}
\title{Introducir publicaciones}
\usage{
introducir_publicaciones(bibfile = NULL, mayus = TRUE, pausa = 3)
}
\arguments{
\item{bibfile}{Ruta al archivo BibTeX (caracter)}

\item{mayus}{Poner todos los caracteres en mayusculas (TRUE) o dejarlos tal cual (FALSE)?}

\item{pausa}{Tiempo (en segundos) que dura el nombre de cada autor en el portapapeles
(para que de tiempo a pegar cada uno en la aplicacion de la ANECA)}
}
\value{
Esta funcion va copiando los distintos campos al portapapeles para
poder pegarlos en la aplicacion de la ANECA.
}
\description{
Introducir publicaciones
}
