% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfs.R
\name{pdf_combinar}
\alias{pdf_combinar}
\title{Combinar varios archivos en un único archivo PDF}
\usage{
pdf_combinar(archivos = NULL, pdf.out = "pdfs_agrupados.pdf")
}
\arguments{
\item{archivos}{Archivos PDF a combinar}

\item{pdf.out}{Nombre del archivo PDF resultante}
}
\value{
Archivo PDF en disco
}
\description{
Combinar varios archivos en un único archivo PDF
}
