% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfs.R
\name{pdf_comprimir}
\alias{pdf_comprimir}
\title{Comprimir archivo PDF}
\usage{
pdf_comprimir(archivo = NULL, calidad = "baja", ...)
}
\arguments{
\item{archivo}{Ruta de archivo PDF en disco}

\item{calidad}{"baja", "media" o "alta" según el grado de compresión deseado}

\item{...}{(opcional) Argumentos extra para \code{\link[tools:compactPDF]{tools::compactPDF()}}}
}
\value{
Fichero PDF comprimido en disco
}
\description{
Comprimir archivo PDF
}
