% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfs.R
\name{pdf_extraer_pags}
\alias{pdf_extraer_pags}
\title{Extraer páginas concretas de archivos PDF}
\usage{
pdf_extraer_pags(
  pdf.in = NULL,
  pags = c("ini", "fin"),
  out.dir = "PDF_SUBSET",
  pdf.out = NULL
)
}
\arguments{
\item{pdf.in}{Nombre del archivo (o archivos) pdf cuyas páginas se desean extraer}

\item{pags}{Vector numérico especificando las páginas a extraer.
Alternativamente, puede indicarse "ini" y/o "fin" para extraer las páginas
inicial y final, respectivamente (ver ejemplos).}

\item{out.dir}{Directorio donde guardar el archivo pdf resultante}

\item{pdf.out}{Nombre del fichero pdf resultante. Si no se especifica,
se utilizará el nombre del archivo original añadiéndole el número de las
páginas extraídas.}
}
\value{
Fichero pdf en disco. Además, la función devuelve las rutas a los archivos extraidos,
por lo que es fácil ejecutar a continuación \code{pdf_combinar()}.
}
\description{
Extraer páginas concretas de archivos PDF
}
\examples{
\dontrun{
pdf_extraer_pags("articulo.pdf")  # extrae página inicial y final por defecto
pdf_extraer_pags("articulo.pdf", pags = "ini")  # extrae solo página inicial
pdf_extraer_pags("articulo.pdf", pags = "fin")  # extrae solo página final
pdf_extraer_pags("articulo.pdf", pags = c(1, 2, 8))  # extrae páginas 1, 2 y 8

## Extraer página inicial y final de muchos pdf a la vez
archivos <- list.files("carpetapdfs", full.names = TRUE)
pdf_extraer_pags(archivos)
}
}
