% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autores.R
\name{pegar_autores}
\alias{pegar_autores}
\title{Pegar autores}
\usage{
pegar_autores(autores = NULL, pausa = 3)
}
\arguments{
\item{autores}{Vector con los nombres de los autores}

\item{pausa}{Tiempo (en segundos) que dura el nombre de cada autor en el portapapeles
(para que dé tiempo a pegar cada uno en la aplicación de la ANECA)}
}
\value{
Los autores son copiados al portapapeles secuencialmente
}
\description{
Dado un vector de caracteres, esta función copiará cada item al portapapeles
secuencialmente. Pude utilizarse para rellenar cualquier campo con múltiples items,
aparte de los autores.
}
\examples{
\dontrun{
autores <- extraer_autores("Quintero, Elena; Rodriguez-Sanchez, Francisco; Jordano, Pedro")
pegar_autores(autores)
}
}
