% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto.R
\name{texto_mayus}
\alias{texto_mayus}
\title{Cambiar mayúsculas de un texto}
\usage{
texto_mayus(
  texto = NULL,
  mayus = c("mayus", "minus", "titulo", "frase"),
  locale = "en"
)
}
\arguments{
\item{texto}{Texto a modificar. Si no se aporta, se leerá directamente
el contenido del portapapeles.}

\item{mayus}{Para cambiar todo el texto a mayúsculas, usar "mayus".
Para cambiar todo el texto a minúsculas, usar "minus".
Para poner en mayúscula sólo la primera letra de cada palabra, usar "titulo".
Para poner mayúscula sólo las primeras palabras de cada frase, usar "frase".}

\item{locale}{Idioma del texto (inglés por defecto, locale = "en"). Para español,
usar locale = "es". Véase \code{\link[stringr:case]{stringr::str_to_sentence()}}.}
}
\value{
El texto reformateado será copiado automáticamente al portapapeles además
de mostrarse en la consola de R.
}
\description{
Cambiar mayúsculas de un texto
}
