% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto.R
\name{texto_unir_lineas}
\alias{texto_unir_lineas}
\title{Unir líneas de texto en una sola}
\usage{
texto_unir_lineas(texto = NULL, mayus = FALSE)
}
\arguments{
\item{texto}{Texto a unir en una sola línea. Si no se aporta, se leerá directamente
el contenido del portapapeles.}

\item{mayus}{¿Convertir todos los caracteres a mayúsculas (TRUE) o dejarlos tal cual (FALSE)?}
}
\value{
El texto reformateado será copiado automáticamente al portapapeles además
de mostrarse en la consola de R.
}
\description{
Esta función une varias líneas de texto en una sola. Puede ser útil, por ejemplo,
para unir texto (títulos, autores, etc) copiado de archivos pdf.
}
