% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicaciones.R
\name{introducir_publicaciones}
\alias{introducir_publicaciones}
\title{Introducir publicaciones}
\usage{
introducir_publicaciones(bibfile = NULL, mayus = TRUE, pausa = 4)
}
\arguments{
\item{bibfile}{Ruta al archivo BibTeX (caracter)}

\item{mayus}{¿Poner todos los caracteres en mayúsculas (TRUE) o dejarlos tal cual (FALSE)?}

\item{pausa}{Tiempo (en segundos) que dura el nombre de cada autor en el portapapeles
(para que dé tiempo a pegar cada uno en la aplicación de la ANECA)}
}
\value{
Esta función va copiando los distintos campos (autores, título, revista, etc)
al portapapeles para poder pegarlos en la aplicación de la ANECA.
}
\description{
Introducir publicaciones
}
\examples{
\dontrun{
bibfile <- system.file("extdata", "articulo.bib", package = "ANECAtools")
introducir_publicaciones(bibfile)
}
}
\seealso{
\code{\link[=extraer_autores]{extraer_autores()}} \code{\link[=pegar_autores]{pegar_autores()}}
}
