% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfs.R
\name{pdf_comprimir}
\alias{pdf_comprimir}
\title{Comprimir archivo PDF}
\usage{
pdf_comprimir(
  pdf.in = NULL,
  pdf.out = NULL,
  calidad = c("baja", "media", "alta"),
  ...
)
}
\arguments{
\item{pdf.in}{Nombre del archivo PDF a comprimir. Si no se especifica, aparecerá
un menú interactivo para elegir el archivo en disco.}

\item{pdf.out}{Opcional. Nombre del archivo PDF resultante. Si no se especifica,
se añadirá el sufijo "_comprimido" al nombre original del PDF.}

\item{calidad}{"baja", "media" o "alta" según el grado de compresión deseado}

\item{...}{(opcional) Argumentos extra para \code{\link[tools:compactPDF]{tools::compactPDF()}}}
}
\value{
Fichero PDF comprimido en disco
}
\description{
Comprimir archivo PDF
}
